My first React Project
Nothing that exciting going on here, I'm just importing React inline as opposed to installing it.

I've even done this with Babbel (Naughty!! Not recommended). I'll go with Node later on and make it a proper setup.
Meanwhile, I'm just running it through apache!

I've just followed some tutorial code for the most part.

I then took a "Clock" project and decided to adapt it into a timer with controls. That's all I've got so far..