Moved my code boxes into a separate function component
This streamlines the code a little but more importantly it gave me an opportunity to consider more asynchronous issues.
Fetch would pull the data after it had rendered the component. Furthermore even if this hadn't been the case, there was the risk of an infinite loop as the DOM would keep changing.

The useEffect hook solves the loop problem and the useState hook allows me to store the code in a state variable. When the state variable is changed by fetchm the DOM is re-rendered.

This exercise taught me the syntactic sugar of class based components is indeed sweet. They are easier to handle IMHO